moved auth configuration to htmlRoutes file
